+++ app/controllers/designs_controller.rb
@@ -535,7 +535,7 @@ class DesignsController < ApplicationController
     end
 
     def generate_import
-      rake_task = "#{RAKE_PATH} design_import DESIGN_ID=#{@design.id} &"
+      rake_task = "#{RAKE_PATH} design_import DESIGN_ID=#{@design.id} SITE_ID=#{params[:site_id].to_i} SUBJECT_STATUS=#{Subject::STATUS.flatten.include?(params[:subject_status].to_s) ? params[:subject_status].to_s : 'pending' } &"
       systemu rake_task unless Rails.env.test?
     end
 end